+++ Car_Proj/ServoMotor.h
@@ -4,6 +4,6 @@ void servo_2right(int duty); // change this to accept pwm duty (angle) as parame
 
 void servo_2center(void);
 
-void servo_turn(float duty);
+void servo_turn(double duty);
 
 void servo_init(void);
